fix: detect removed fields without nil deref error
If the property being evaluated during a diff does not exist in the new
schema, we cannot make a copy of that that property's nested schema as
it would result in a null pointer dereference. This is handled by
assigning the new schema in the diff as empty.

The change here simply ends the diff determination for the given
property at this point, rather than attempting to further act on the
nil schema.