Update README to include pronunciation of 'crdify'
It became clear that folks were not understanding how we meant this project name to be pronounced. It's almost like "certify" but with a d in the middle.

According to google, certify is `sur·tuh·fai` so I figure `sur·duh·fai` represents the intended pronounciation of the project name.